Refactoring : slideContent와 slideNavi 두 클래스 병합 index를 저장하지 않기 위해 setIndex메서드 제거
