+++ src/js/view/slide/SlideContent.js
@@ -1,35 +1,41 @@
-import { checkLocalItem, getLocalItem } from '../../util.js';
 import { slideTpl } from '../../template/slideTpl.js';
-
+import { throttle } from '../../util.js'
 export default class SlideContent {
-    constructor({ slideListEl, urlRequestData }) {
+    constructor({ slideListEl, naviEl }) {
         this.slideListEl = slideListEl;
-        this.currentIdx = 0;
-        this.init(urlRequestData);
+        this.naviEl = naviEl;
     }
 
-    async init(url) {
-        if (await checkLocalItem(url)) return;
-        this.setContentData(getLocalItem(url));
+    init(idx) {
         const currentContentEl = this.slideListEl.children[1];
-        this.render(currentContentEl, this.getCurrentContent());
+        this.render(currentContentEl, this.getCurrentContent(idx));
     }
 
-    showPreContent() {
+    clickNextBtn(handler, throttleTime) {
+        const nextBtn = this.naviEl.querySelector('.slide-next');
+        nextBtn.addEventListener('click', throttle(handler, throttleTime))
+    }
+
+    clickPreBtn(handler, throttleTime) {
+        const preBtn = this.naviEl.querySelector('.slide-prev');
+        preBtn.addEventListener('click', throttle(handler, throttleTime))
+    }
+
+    showPreContent(idx) {
         const preContentEl = this.slideListEl.children[0];
         const nextContentEl = this.slideListEl.children[2];
 
-        this.render(preContentEl, this.getCurrentContent());
+        this.render(preContentEl, this.getCurrentContent(idx));
 
         this.slideListEl.insertBefore(nextContentEl, preContentEl);
     }
 
-    showNextContent() {
+    showNextContent(idx) {
         const preContentEl = this.slideListEl.children[0];
         const nextContentEl = this.slideListEl.children[2];
 
         //그 다음에 보여줄 content를 render한다.
-        this.render(nextContentEl, this.getCurrentContent());
+        this.render(nextContentEl, this.getCurrentContent(idx));
         //맨앞에 있는 li엘리먼트를 맨 마지막 순서로 옮겨준다. 
         this.slideListEl.appendChild(preContentEl);
     }
@@ -42,17 +48,7 @@ export default class SlideContent {
         this.contentData = contentData;
     }
 
-    setCurrentIdx(index) {
-        this.currentIdx = index;
-    }
-
-    getCurrentIdx() {
-        const contentCount = this.contentData.length;
-        this.currentIdx = this.currentIdx < 0 ? contentCount - 1 : this.currentIdx % contentCount;
-        return this.currentIdx;
-    }
-
-    getCurrentContent() {
-        return this.contentData[this.getCurrentIdx()];
+    getCurrentContent(idx) {
+        return this.contentData[idx];
     }
 }
